move `create_event` to `confetst` to reduce duplidcation in test.
